Cleaned functions around main parser function #75
